backend/requirements.txt
# Core Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.10.0

# LangChain Ecosystem
langchain==0.3.19
langchain-core>=0.3.0
langchain-google-genai==2.0.10
langchain-groq==0.2.1
langgraph==0.2.64
langgraph-checkpoint==2.0.10
langgraph-checkpoint-postgres>=3.0.0

# PostgreSQL
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0

# LangGraph Store (Long-term Memory)
langgraph-store-postgres>=0.1.0

# Deep Agents Framework
deepagents>=0.1.0

# Vector Store (PGVector + Embeddings)
langchain-postgres>=0.0.16
langchain-google-genai>=2.0.10

# Search and Web Tools
tavily-python==0.5.0
arxiv==2.2.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# Async HTTP for parallel verification
aiohttp>=3.9.0

# Document Export
python-docx==1.1.2
reportlab==4.2.5
pypdf2==3.0.1
pypdf>=5.0.0

# Utilities
uuid7==0.1.0
python-multipart==0.0.20

# AI/ML Models
google-generativeai==0.8.3

# GitHub Integration
PyGithub>=2.1.0
